+++ src/settings.ts
@@ -2,13 +2,15 @@ import path from 'path';
 import {cosmiconfig} from 'cosmiconfig';
 import defaults from 'defaults';
 import {BranchObject} from 'semantic-release';
+import fs from 'fs/promises';
+import micromatch from 'micromatch';
 
 const CONFIG_NAME = 'workspaceRelease';
 
 export const SETTINGS = {
   preReleaseVersionTemplate: '${version}',
   registry: 'https://registry.npmjs.org',
-  workspace: path.join(process.cwd(), 'packages'),
+  workspaces: ['packages'],
   tagFormat: '${name}@${version}',
   release: {
     extends: 'semantic-release-commit-filter',
@@ -30,28 +32,68 @@ export const SETTINGS = {
   extendsReleaseRules: [],
   extendsNoteGeneratorTypes: [],
   extendsDefaultPlugins: [],
+  perPackageConfig: {},
+  autoCreatedPlugins: false,
 };
 
 export async function importSettings() {
   const explorer = cosmiconfig(CONFIG_NAME);
   const result = await explorer.search();
 
+  try {
+    const workspacePackageJsonPath = path.join(
+      result.filepath,
+      '..',
+      'package.json',
+    );
+    const {workspaces} = JSON.parse(
+      await fs.readFile(workspacePackageJsonPath, 'utf8'),
+    );
+    SETTINGS.workspaces = workspaces;
+  } catch (error) {
+    console.warn('Could not read workspaces from package.json');
+  }
+
   if (result) {
     Object.assign(SETTINGS, defaults(result.config, SETTINGS));
     SETTINGS.release.plugins = result.config?.release?.plugins;
   }
 
-  SETTINGS.release.plugins ??= defaultPlugins();
+  if (!SETTINGS.release.plugins) {
+    SETTINGS.release.plugins = createPlugins();
+    SETTINGS.autoCreatedPlugins = true;
+  }
 }
 
-function defaultPlugins() {
+export function generatePackageSettings(packageLocation: string) {
+  const pattern = Object.keys(SETTINGS.perPackageConfig).find(pattern =>
+    micromatch.isMatch(packageLocation, pattern),
+  );
+  const customConfig = SETTINGS.perPackageConfig[pattern];
+
+  if (!customConfig) {
+    return SETTINGS;
+  }
+
+  const newConfig = structuredClone(SETTINGS);
+  Object.assign(newConfig, customConfig);
+  newConfig.release = Object.assign({}, SETTINGS.release, customConfig.release);
+
+  if (newConfig.autoCreatedPlugins && !customConfig?.release?.plugins) {
+    newConfig.release.plugins = createPlugins(newConfig);
+  }
+
+  return newConfig;
+}
+
+function createPlugins(settings = SETTINGS) {
   return [
     [
       '@semantic-release/commit-analyzer',
       {
         preset: 'angular',
         releaseRules: [
-          ...SETTINGS.extendsReleaseRules,
+          ...settings.extendsReleaseRules,
           {
             breaking: true,
             release: 'major',
@@ -68,7 +110,7 @@ function defaultPlugins() {
             type: 'breaking',
             release: 'major',
           },
-          ...(SETTINGS.npmRelease
+          ...(settings.npmRelease
             ? [
                 {
                   type: 'docs',
@@ -145,15 +187,15 @@ function defaultPlugins() {
         },
       },
     ],
-    ...(SETTINGS.changelogCommit ? ['@semantic-release/changelog'] : []),
-    ...(SETTINGS.npmRelease ? ['@semantic-release/npm'] : []),
-    ...(SETTINGS.changelogCommit
+    ...(settings.changelogCommit ? ['@semantic-release/changelog'] : []),
+    ...(settings.npmRelease ? ['@semantic-release/npm'] : []),
+    ...(settings.changelogCommit
       ? [
           '@semantic-release/github',
           ['@semantic-release/git', {assets: ['CHANGELOG.md', 'LICENSE']}],
         ]
       : []),
-    ...SETTINGS.extendsDefaultPlugins,
+    ...settings.extendsDefaultPlugins,
     [
       '@semantic-release/exec',
       {